crash: Fix warning
* trans/crash.c (template_make_file_name): Fix format for time_t.

+++ trans/crash.c
@@ -149,7 +149,7 @@ template_make_file_name (const char *template,
 	      break;
 
 	    case 't':
-	      fprintf (stream, "%d", time (NULL));
+	      fprintf (stream, "%lld", (long long) time (NULL));
 	      break;
 
 	    default:
